backend/requirements.txt
fastapi==0.115.12
fitz==0.0.1.dev2
openai==0.27.10
pdf2image==1.17.0
pdfplumber==0.11.6
pinecone==7.0.1
pydantic==2.11.5
pydantic_settings==2.9.1
pytesseract==0.3.13
python-dotenv==1.1.0
